Ya he corregido el tipo de dato para que la entidad Register maneje decimales en sus atributos humidity y temperature.

+++ src/main/java/com/example/api_plantanciones/dto/plantation/PlantationAvgHumTemByDate.java
@@ -1,4 +1,4 @@
 package com.example.api_plantanciones.dto.plantation;
 
-public record PlantationAvgHumTemByDate(Long idPlantacion, Long avgHumidity, Long avgTemperature) {
+public record PlantationAvgHumTemByDate(Long idPlantacion, Double avgHumidity, Double avgTemperature) {
 }
